refactor: give the Instance History sort order a single home for its parts
Review round two. Four structural notes, all pointing at the same thing:
pieces of the sort-order feature had been scattered wider than their reach.

The tree store redeclared its own `DEFAULT_SORT_ORDER: 'desc'`, a second
copy of the value that `stores/instanceHistorySortOrder` already owns. Two
literals with no link between them drift the moment one is changed, so the
store now exports `DEFAULT_ORDER` and the tree store imports it. The field
stays a fallback that `setRootNode` overwrites before any request is built,
which is why the exported default is used rather than `INITIAL_ORDER` -
binding a reset value to whatever localStorage held at module load adds
coupling for a value that is never read.

`buildElementInstanceSort` sat in `modules/utils`, a location that implies
app-wide reach, while all four call sites are inside ElementInstanceLog. It
moves next to them so its scope is legible from its path.

`Layout` had grown a `SearchRow` around its own `searchInput` slot plus a
hardcoded `SortOrderToggle`, so the caller no longer controlled what it
passed. The row is assembled at the call site instead and `Layout` goes back
to rendering the slot it is given. The gate that kept a lone sort toggle off
skeleton and error panels becomes structural - those branches simply pass no
slot.

`SearchRow` also loses its `padding-right`. Carbon's ghost button already
reserves 15px inline-end, so the extra 8px pushed the control's label to
23px from the panel edge while the header toggles above sit at PanelHeader's
16px. Dropping it lines the two up to within a pixel, using Carbon's own
spacing instead of an added rule.
