+++ operate/client/src/App/ProcessInstance/ElementInstanceLog/index.tsx
@@ -56,17 +56,7 @@ const Layout: React.FC<LayoutProps> = observer(
             )}
           </PanelHeader>
         )}
-        {/* Gated on searchInput as well as the mode, because the skeleton and
-            error branches render Layout without one - otherwise this row shows a
-            lone sort toggle over a panel that has no tree to sort. The header
-            toggles above keep rendering in those states; that is pre-existing
-            and deliberately left alone here. */}
-        {!modificationsStore.isModificationModeEnabled && searchInput && (
-          <SearchRow>
-            {searchInput}
-            <SortOrderToggle />
-          </SearchRow>
-        )}
+        {!modificationsStore.isModificationModeEnabled && searchInput}
         {children}
       </Container>
     );
@@ -139,7 +129,12 @@ const ElementInstanceLog: React.FC<{isPanel?: boolean; showHeader?: boolean}> =
       <Layout
         isPanel={isPanel}
         showHeader={showHeader}
-        searchInput={<SearchForm />}
+        searchInput={
+          <SearchRow>
+            <SearchForm />
+            <SortOrderToggle />
+          </SearchRow>
+        }
       >
         <PanelBody>
           <ErrorBoundary
