+++ operate/client/src/modules/stores/instanceHistorySortOrder.ts
@@ -62,4 +62,4 @@ class InstanceHistorySortOrder {
 
 const instanceHistorySortOrderStore = new InstanceHistorySortOrder();
 
-export {instanceHistorySortOrderStore};
+export {instanceHistorySortOrderStore, DEFAULT_ORDER};
